squash changes to add input building to zkasper_cli
