+++ justfile
@@ -2,7 +2,7 @@ MAINNET_ELECTRA_SLOT := "11649026" # two epochs after the fork, otherwise we wil
 SEPOLIA_ELECTRA_SLOT := "7118912" # two epochs after the fork, otherwise we will need pre-fork state to process attestations
 
 test-sync network slot:
-    cargo run -r --bin host --no-default-features --features {{network}} -- --network {{network}} sync --start-slot {{slot}} --log-path ./logs/{{network}}.log --mode ssz
+    cargo run -r --bin zkasper_cli --no-default-features --features {{network}} -- --network {{network}} sync --start-slot {{slot}} --log-path ./logs/{{network}}.log --mode ssz
 
 test-sync-full-mainnet:
     @echo "Testing sync on mainnet from slot $(MAINNET_ELECTRA_SLOT)"
